Windows: Factoring out unused fields
