+++ daemon/inspect.go
@@ -21,51 +21,9 @@ func (daemon *Daemon) ContainerInspect(name string) (*types.ContainerJSON, error
 		return nil, err
 	}
 
-	mountPoints := make([]types.MountPoint, 0, len(container.MountPoints))
-	for _, m := range container.MountPoints {
-		mountPoints = append(mountPoints, types.MountPoint{
-			Name:        m.Name,
-			Source:      m.Path(),
-			Destination: m.Destination,
-			Driver:      m.Driver,
-			Mode:        m.Mode,
-			RW:          m.RW,
-		})
-	}
+	mountPoints := addMountPoints(container)
 
 	return &types.ContainerJSON{base, mountPoints, container.Config}, nil
-}
-
-func (daemon *Daemon) ContainerInspectPre120(name string) (*types.ContainerJSONPre120, error) {
-	container, err := daemon.Get(name)
-	if err != nil {
-		return nil, err
-	}
-
-	container.Lock()
-	defer container.Unlock()
-
-	base, err := daemon.getInspectData(container)
-	if err != nil {
-		return nil, err
-	}
-
-	volumes := make(map[string]string)
-	volumesRW := make(map[string]bool)
-	for _, m := range container.MountPoints {
-		volumes[m.Destination] = m.Path()
-		volumesRW[m.Destination] = m.RW
-	}
-
-	config := &types.ContainerConfig{
-		container.Config,
-		container.hostConfig.Memory,
-		container.hostConfig.MemorySwap,
-		container.hostConfig.CPUShares,
-		container.hostConfig.CpusetCpus,
-	}
-
-	return &types.ContainerJSONPre120{base, volumes, volumesRW, config}, nil
 }
 
 func (daemon *Daemon) getInspectData(container *Container) (*types.ContainerJSONBase, error) {
@@ -104,9 +62,6 @@ func (daemon *Daemon) getInspectData(container *Container) (*types.ContainerJSON
 		State:           containerState,
 		Image:           container.ImageID,
 		NetworkSettings: container.NetworkSettings,
-		ResolvConfPath:  container.ResolvConfPath,
-		HostnamePath:    container.HostnamePath,
-		HostsPath:       container.HostsPath,
 		LogPath:         container.LogPath,
 		Name:            container.Name,
 		RestartCount:    container.RestartCount,
@@ -114,10 +69,12 @@ func (daemon *Daemon) getInspectData(container *Container) (*types.ContainerJSON
 		ExecDriver:      container.ExecDriver,
 		MountLabel:      container.MountLabel,
 		ProcessLabel:    container.ProcessLabel,
-		AppArmorProfile: container.AppArmorProfile,
 		ExecIDs:         container.GetExecIDs(),
 		HostConfig:      &hostConfig,
 	}
+
+	// Now set any platform-specific fields
+	contJSONBase = setPlatformSpecificContainerFields(container, contJSONBase)
 
 	contJSONBase.GraphDriver.Name = container.Driver
 	graphDriverData, err := daemon.driver.GetMetadata(container.ID)
